refactor(perry-ext-net): split raw-net bridge into raw_bridge.rs (file-size gate)
lib.rs exceeded the 2000-line lint gate after the #2154 raw-mode additions.
Move RawReadState, the raw_state_for/raw_mark_closed helpers, the four
perry_net_raw_* C-ABI vtable fns, and the vtable registration into a new
raw_bridge submodule, and route run_socket_task's raw handling through two
thin hooks (route_data / mark_terminal). Behavior-identical; lib.rs back to
2000 lines.
